requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
confuse==1.3.0
croniter==0.3.34
dask==2.28.0
distributed==2.28.0
docker==4.3.1
HeapDict==1.0.1
idna==2.10
immutables==0.14
marshmallow==3.8.0
marshmallow-oneofschema==2.0.1
msgpack==1.0.0
mypy-extensions==0.4.3
natsort==7.0.1
pendulum==2.1.2
prefect==0.13.8
psutil==5.7.2
psycopg2-binary==2.8.6
python-box==5.1.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sodapy==2.1.0
sortedcontainers==2.2.2
tabulate==0.8.7
tblib==1.7.0
text-unidecode==1.3
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
urllib3==1.25.10
websocket-client==0.57.0
zict==2.0.0
